ai_plugins/requirements_yolo.txt
fastapi
uvicorn[standard]
ultralytics
opencv-python-headless
numpy
pillow
httpx
